Add test for running the spec test cases
Adds the source map tests repo as a submodule, and also adds a test file that
will run all of the test cases in the repo. It only checks the valid/invalid
flag of the test cases and does not check other test actions.

Some tests are skipped due to limitations that aren't easily worked around
at the moment.

Also fix some small issues related to sources/sourcesContent that needed
fixing to pass the non-skipped tests.