Refine handling of null sources in various places
When a source is null, it won't be included in the file map and checks that use
a file won't be run. In the case that a sourcesContent is provided, it would
potentially be better to treat that as an anonymous file but we currently don't.

Also check for the string "null" in some cases, as the underlying source maps
library can return this instead of a null.